Method reflection refactorings.
